Fix display of recurring in thank you preview
To test create a recurring contribution in CiviCRM and attempt
to send the contact a thank you from the action on the right of
their contribution (on their contribution tab). The
text in the preview does not detect this
is a recurring contribution, absent this patch.

Per the phab only the preview (not the send)
is affected

Bug: T342081
Change-Id: Ie8be832e46eaabd0ce412c082f51e89f4d95df16