Follow up fixes on snooze
This moves the queueing of the snooze to after
the email is saved - it's all a bit tricksie but basically we were never
gonna be confident the contact had already been created when
listening to the pre hook.

The hook_custom is called when the snooze value has been updated -
so there is definitely a contact ID at that point. Fixing that
allowed some other minor fixes

https://docs.civicrm.org/dev/en/latest/hooks/hook_civicrm_custom/

Bug: T349277
Change-Id: Iea35b9d7374ed0f05ee127e53fbfa34aae1e2278