Remove upgrade to mailing_event_queue table
We already made this change on prod.

Note we did fill the mailing_id field on create - but have
not been populating it since then so we
see

 select count(*) FROM civicrm_mailing_event_queue WHERE mailing_id IS NULL LIMIT 50;
+----------+
| count(*) |
+----------+
|     3313 |

We need to make sure we are populating it going forwards as a follow up (rolling
out 5.67 may or may not be enough of itself?

Bug: T350208
Change-Id: I6fbc5374c0fd8b623878530837275a1f1b0f3804