Add last donation amount on summary profile
Since we only have the last amount and currency on the wmf_donor
table for the annual fund (not endowment) we need to use a custom
hook to get the last all_funds amount and currency for the
summary profile. We add it whenever we find the
all_funds_last_donation_date field in a profile. So to test this,
go to Administer->Customize Data and Screens->Profiles then click
the 'Fields' section of 'Summary Overlay' and add the
'All Funds Last donation date' field. Then go to a contact with
a recent donation and hover over the icon before the name to see
the overlay. Check that a new 'Last donation amount' field is
also present with the amount and currency.

Bug: T346917
Change-Id: I15bacf6dab983cc090379b53163d6cad5fda1ef0