Set outcome to retry to avoid the queue being set to aborted
If the outcome is set to abort then the
calling code updates the queue to 'aborted'.

Changing the status to 'retry' allows it to be released without changing the
queue.

I have mixed feelings about this as I suspect it might be more
correct to set it to delete & allow the calling function to delete it,
while the hook could create a new item in the damaged
queue rather than to update the existing

Bug: T350680

Change-Id: Icf1defec81f0a142509ecd81458eaeb285a5cd1c