Fix authorization problem on getUpgradeable
It turns out the fail really was inadequate api permissions. This appeared to work
on dev because we were testing in a browser with a logged in user. In a private browser window
the same erro as on prod occurs.

I would note that the api level is probably the right place for permission checking so I put it
there - but I think it should be removed from
\Civi\WMFHelpers\ContributionRecur::getUpgradeable() and the PreferencesLink code
should either not pass the check sum or should call the api.

Change-Id: I33988ee6718efdd8eae4e5fb20df11f860a8073c