Fix recurring record reactivation
Make it actually change the status, not just blank the cancel/end
dates & cancel reason. Also add a guard to only reactivate records
when the recurring charge schedule is handled by the processor,
to avoid reactivating something that we control and has just been
cancelled by donor relations.

I was going to use the getDefaultPaymentProcessor function added
in the latest SmashPig version, but instantiating the class seems
to be a bit troublesome.

Bug: T354357
Change-Id: I5ac7f65fbed26fe15be5cc40e91f680c391a7fb7