Fix Benevity import to use configured source data
The value Nora wants in the Campaign/Gift Source field is to be derived from the import file.

The value is not always the same as the mapping file value and Nora has provided a mapping.
In line with our practice of moving stuff to the extension when we touch it I figured this
mapping should live in the extension & be transformed there - hence
I added the functions to the import hook class, which is where I expect them
to eventually be called from. I feel like there might be a place for
user-configurable mappings going forwards, hence I put it in a json file to lay
a trail for this approach.

There is still a bit of logic in the Benevity import so we can't quite flip the
switch

Bug: T354911
Change-Id: I0743f126026ac050c21a3412a2952b7c5f4e3f30