Turn fredge tables into view-based entities
This creates views for the fredge tables which CiviCRM
can interact with as if they were in the CiviCRM database,
allowing us to add create,delete api to use from our
AntiFraudQueueConsumer.

I had to move the table creation from drupal to the extension
in order to get the tables to install first. The
removed file only relates to install & uninstall
https://api.drupal.org/api/drupal/modules%21system%21system.api.php/function/hook_schema/7.x

We had a hack in place that was enough to get Get to work but
it fell down on save & update.

Note that I tested on prod & the civicrm user can create tables so this
create view will work under their context.

The only live code affected is the use of the fredge api in
forgetme. The code can be stepped through in the forget me
test but note that what you see is the api not breaking rather than
the test passing. That test fails for me before and after this
change (and it turns out the extension is not wired into
phpunit.xml). I can look at that as a follow up.

Further notes on naming
1) I renamed Wmf to WMF inline with other changes we have been making
2) It would be consistent with our recent renaming to
switch from PaymentsFraud to PaymentFraud but on the other
hand the fredge table is payments_fraud. I am open to doing a rename
on the basis that the Civi table, DAO, and api would be consisent with
each other and not with the fredge table - but if we do that are we
happy with 'PaymentInitial'? (I find that name a bit confusing but
might be alone in that).
3) It would be consistent with recent efforts to rename the
extension 'wmf-fraud'. I'm open to doing that

Bug: T356115

Change-Id: I2a5474318c902458eab3ff7588889740ba4f3dc6