Fix default financial_type_id for recurring
Hard-code to Cash to match RecurringQueueConsumer. Old code was
mistakenly sending the numeric value from the $msg as a :name,
but we don't even want to copy the financial_type_id from the
contribution row to the contribution_recur row. For one thing,
it changes from the first to second installment.

Bug: T356999
Change-Id: I9f20c4887ae0877b9a49e969dd7c07a91b5c8e24