Check Civi contribution tracking not drupal from wmfImportTest
This test is probably close to going since I'm not sure the import it tests is
really used now, or it won't be for long. I switched it to the Civi table & also
removed the call to the custom date function. I don't know for sure whether it
resolves to something sane - but the test passes so I assume it's not material to the
test & it's one shuffle towards the end of that function

Bug: T356115
Bug: T34708
Change-Id: I64ff783ad6750757791945bf8e1ba8f17808e7a0