Fix update of dedupe listing after merge
Be warned - it's easy to not load the new js when applying this patch - refresh in place
didn't always seem to do it for me.

A good way to set up for this is to create 3 mice in the browser console

CRM.api3('Contact', 'create', {'first_name' : 'Mickey', 'last_name' : 'Mouse', 'contact_type' : Individual, 'email' : 'm@example.com'} );
CRM.api3('Contact', 'create', {'first_name' : 'Mickey', 'last_name' : 'Mouse', 'contact_type' : Individual, 'nick_name' : 'Mice', 'email' : 'm@example.com'} );
CRM.api3('Contact', 'create', {'first_name' : 'Mickey', 'last_name' : 'Mouse', 'contact_type' : Individual, 'email' : 'm@example.com'} );

Load up deduper & use nick_name = 'mice' as the filter.

If you do this with no pre-exisitng mice you should get presented with 2 merge tiles. In one
case the contact with the nick name will be on the left & on the other the right. Take note
of the ID of this mouse. Merge the tile where the nick-named mouse is on the left. On master
the other merge will disappear too.

This is because the contact on the left hand side was deleted & hence cannot be used for
further matches.

However, the patch changes it such that the dedupe tiles
are retained and updated to the id of the contact that is retained. If doing so
would create a duplicate the match is removed.

Bug; T339288

Change-Id: I4f30cfa939677d235f8da618779c25ca1addb57a