Move converstion rate handling to getters
Note this surfaced an inconcistency in the rounding of the original
amount as locked in by the tests in fundraiseUpAudit versus those
tests edited in this patch

This inconsistency exists in production - eg. scroll down the contributions
here
https://civicrm.wikimedia.org/civicrm/contact/view?reset=1&cid=11749254&selectedChild=contribute
and note some are 'USD 10.4' and others are 'USD 10.40'. I consolidated
on 'rounding-to-currency'

Bug: T357471
Change-Id: I68a3d7f1fd7411628de79c187fb79132b8758416