Move data wranging for fredge to extension
This commissions the fredge functions that declare the drupal table schema & instead uses a FraudMessage object
to do the wrangling on the 3 fields which were being altered in the prep_data function.

I went back & forth on this - ie each of the 2 functions that call the
message class only do fixes on 2 fields so there was a case to just do that in-function
rather than have a whole nother class to handle it. The 2 functions are also
dealing with different messages so that kinda strenghtens the case
for them each just handling their own message. However, there is
also a case for having all our message wranging in
the WMFMessage namespace

Also, I followed back the data formatting functions & while they arguably handled
date time zones it didn't seem to me like they really did anything different to
the simpler date() call I used. The constructor of the Date object says

'The $timezone parameter and the current timezone are ignored when the $time parameter either is a UNIX timestamp (e.g. @946684800)'

Which is the context in which it is constructed here

Bug: T357469
Change-Id: Id9d844bb41bf5df593cb946ab7d258784fe5ccff