Move OptInTest to the extension
I went through & updated all api calls to apiv4 where possible.
I also used the new api to wrap the queue processing & consolidated some
places in the moved classes to use that.

Note that the parent class is implementing the transactional inteface.
I have mixed feelings about that interface

good
- it potentially performs better
- it uses mysql rollback & so cleans up fully

bad
- it has to be abandoned as soon as we do
any query in a test that does not work
with mysql transactions (eg create a table)
- you can't see the current db state from outside
the test as the transactions are not committed

(Note any switch away from it would be outside the
scope of this patch)

Bug: T357469
Change-Id: I128cdca585850e27625ab9128f76db2519e4b835