Return handling for tags & groups to import classes
These fields are only set in the import classes & in one case (contribution_tags) in Donation
QueueConsumer. In that case it is an array so needs no handling here. In the imports it
is always a string. Note that the array_unique is handled at the point of save
so we can remove from here

Bug: T357471
Change-Id: I671ce29302d63de2ab7d5026d9c4eba54f9c22db