Use Civi Setting for braintree (only)
This removes usage of the drupal variables for braintree & switches
to CiviCRM settings. Note that the goal is to remove reliance on drupal functions
but also to not have such a crazy quantity of esttings.

The json files wind up nested quite deep in the test suite
due to the switching of directories to imitate the prod directory but
with only 1 file in there. An alternative (but perhaps not worth the effort)
would be to use just 1 directory & move the file into it to run the test

Change-Id: Ib2ff159ecae58b99d3c837da59bbe02024a3d20b