Move a couple more checks to the Message Object
This moves the check for contribution tracking ID to the message object
and keeps the management of the contribution recur ID in sync
by allowing setters in both cases.

This is a subset of code in
https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/1018389
which I haven't quite been able to land and also
https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/1016908

I also brought the call to queue the create back in - it needs to come out of that
to-be-removed code & was trivial so I put it in the main function. I didn't touch
the remaining caller of that code as I'm expecting that
code to rely on this code to create the tracking record.

WHere I'm working towards is simplifying
wmf_civicrm_should_update_contribution_tracking

to being "if not recurring or is recurring and has no prior
contributions" as I think the more complicated logic is more or less that
and where it is less it is in a bad way - ie because it is
skipping things coming in from the recurring donation processor
when they need to be done & hence the recurring donation processor has to do them

Bug: T240581

Change-Id: I4608813044ac5bf1c0786a468f4fa0b9ab0b8029