Move handling of saving contribution_tracking to main processing
This is already done in wmf_civicrm.module for most messages
but it has a convoluted logic which excludes the ones coming from the Recurring queue.

This makes the logic to add the contribution to the contribution tracking in all
cases where the contribution tracking exists & there is not a prior contribution against the
recurring - which I think is what the rest adds up to.

There is also a carve out for matching gifts - that one should be on the way out...

Change-Id: Ia45f929cacac1dceb300785ee8837d95f12f6582