Only call resolveOrganization when there is a soft credit involved
The resolveOrganization function finds the organizationName which is only used when
a soft credit is involved, but if it does not find something it throws an exception.

This adds an IF to only do the check if we are interested in the results.
Since the else-if clause was already checking if the SoftCredit was empty
I moved that to be an outer if around both

Bug: T360150
Change-Id: Icd1a414d535c92709119a180998653e00fe942a9