Establish Contact Update behaviour in test, remove extraneous code
I added a test to clarify what happens when the updateContact is called from
the recurring queue consumer.

I was able to establish that the behaviour is unchanged when I remove
the line that I have deleted. This would not be the case if street_address
were not present but I don't think that's a scenario we need to handle

Change-Id: Ib6cfec0534abbcd72340660728c1689fd75934b0