Consolidate calls to message_import within Benevity
I'm aiming to call only a subset of the stuff that is in the main message_import as
most of it is unrelated. This is a first step, no longer calling it directly
and eliminating the use of pass-by-reference. I'm pretty sure that the only
output used from the pass-by-reference is the contact_id in the first Benevity
call (where it is used to set up the second call)

Bug: T240581

Change-Id: I1fbdf42fee706faec1f1040e9111a8e8f675d0c5