Separate import code from main message_import
Almost all the code in the message_import class is irrelevant to the offline2civicrm
but it adds complexity on both sides having the import code go through it.

This copies back the parts that ARE shared. Hopefully we can stop calling
normalize() as the import class does it's own normalize but it's in for now
By contrast I was pretty confident on validate

This means that we are no longer adding contribution tracking records for these rows
- that is inline with the way we have been moving with contribution tracking.

Some more work to be done before we can unravel that contact handling

Note that once this & tests are not calling message_insert
I think that function will move into the DonationQueue class as that
is really what it relates to

Bug: T240581

Change-Id: I5c1ea3bcbdaf6c0d7ab38ea855c79f6396fe27ad