Only set no-thank-you reason for matchig gift or workplace giving
Per the discussion at
https://phabricator.wikimedia.org/T360150#9720638

our original rule for suppressing thank you emails during imports is too broad,
affecting all imports where the user does not add a column to override it.

This rule no longer serves us in part because our business practice has changed and
we now send thank yous for donations that came in through the portals and in
part because we have extended the range of things going through the import.

Melanie has identified that basing the default on the soft credit type makes more sense.

At this stage Benevity is still going through offline2civicrm and has specific
coding for the no-thankyou. When we move it over we will need to address that -but
the fact it is in the Benevity code should work to remind us.

Bug: T360150
Change-Id: I7c0c7b0c14ac7aa4ba7fcec88a8caa0bd56c88ce