Get rid of name_prefix & name_suffix in favour of the api fields
This seems to me to exist only for these imports - in which case the best
field name is really the api v4 name & then we don't need to map & re-map.

If we DO need to map from an external array key schema to an internal one
we should to that in the DonationMessage() class but I don't think we
do need to do that here.

Note that I've tried to lay the ground work for mapping the custom fields
between apiv3 & v4 to progress switching over

Bug: T240581

Change-Id: Ibcf15434eecffc47fb369b0fd46b760c87faf42d