Update OptInQueueConsuer to use a Message class, apiv4
I'm working towards cleaning up the custom handling mess in WMFContact::save because it keeps breaking my brain

The goal is that the fields would come out of Message->normalize() in apiv4 ready format because
those classes are responsible from mapping our externally facing field names to the ones to
be used by the api.

Currently the Save function uses both apiv4 & v3 & has 2 paths - seemingly 'create' and 'update' but it's
not that clean. I'm trying to get both to use apiv4 & towards that have tried to put the
wrangling between the 2 in its own function in
https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/1022309

The next step on that is custom fields & the obvious custom fields to start with appear to
be opt_in and do_not_solicit as they appear in both create & update - since those come
from the opt-in code as well I decided to tackle getting them to hit the Save function as both
the old fields & the apiv4 ready one so we only need to fix the DonationMessage to able to
add handling.

This also eliminates a function in the legacy code base that validates the hash, moving
it to the message classes and one call to an apiv3 custom field function that we hope to remove

Bug: T357471
Change-Id: I8dd46be207843e5ba37c23e8dc4209efc31607c8