Override core mailing action urls with ones to our preference centre
This replaces the urls generated by

{action.unsubscribeUrl}

{action.optOutUrl}

with our own ones. I used the same one in both places because we don't want
to ask users to remember which is which (I barely can)

These can be seen in the preview part of the screen at

https://wmff.localhost:32353/civicrm/mailing/send?reset=1

Per code comments there is some question as to where the extracted function
should live (which could be a follow up).

Bug: T361630

Change-Id: I4afd51e223487a7c22be2040a0fb26eb087dc129