Move unpacking of external identifiers to message class
The goal is that the Message classes should translate our incoming messages
to CiviCRM fields and that by the time it reaches the WMFContact::save the
values should be interpreted  (hopefully the save will not
do much that Contact:save() does at that point & we can use the latter
more freely).

Moving it to normalize requires that the message passed to ContactSave is
always normalized. This was already the case for most
callers - ie RecurDonation, UPIDonation & Donation
but had to be changed for ModifyAmount.

IF is the case for the imports classes too - although they do not use these fields

I moved one test to the DonationQueue test & ensured the remaining ones used different
normalized messages (ModifyAmount & DonationRecur)

Change-Id: I28110e4e10ed4a4ea2cee43ca4f3fb53613cce13