Copy handleUpdate back to ChecksFile
While duplicating code seems counter intuitive we have a function that is super-confusing, in part
because it has a whole load of code that is import specific. This copies the function back to
the import function, removing stats tracking & external identifier handling
and removes the handling for organization detail from the main function.

Within the main function it exists only for purposes of handling the organizations
imported via checks.

One significantly mitigating feature about this code duplication is that that
code in the import class is getting ever closer to being decommissioned, with the
organization imports having a fairly short road ahead of them - at which point it will all just
go

Change-Id: I0eb725f26210ec0621727c05d0a25e3050e0f5f1