Backfill payment_instrument_id for monthly convert
We need the payment_instrument_id when an autorescue comes in for
contribution_recur rows with no prior contribution. These are all
monthly convert rows, and the subscr_start message that comes in
when we establish them links the contribution_recur to the one-time
contribution by assigning the one-time contribution's
gateway_txn_id to the contribution_recur's processor_id.

Change-Id: Ia1b1818fd47c21d635f58fd694d26a26b3442570