Exchange Rate Convert API action
Has a configurable cache granularity (old was hardcoded to 1 rate
per day). Falls back to hardcoded rates in SmashPig's CurrencyRates
class, but maybe we shouldn't require that library as a dependency.

There was an intermediate fallback step in the old code which I don't
think actually did anything useful.

It looks like there's no way to specify an API4 parameter as a float,
so I had to make the from_amount a string.

I also had to add default values of '' for the required fromAmount
and fromCurrency parameters to avoid seeing these
[PHP Notice] Undefined index: fromAmount at {...}AbstractAction.php:307
[PHP Notice] Undefined index: fromCurrency at {...}AbstractAction.php:307

Bug: T270666
Change-Id: Ic855a345ce3257f1eb4daed039e126062254c035