Remove now unused get/set contactID from WMFContact.save
We have now removed the places that used to call setContactID() from the RecurringQueueConsumer
and ModifyAmountQueueConsumer

That leaves us with just the clean up job - as will be apparent looking at
this patch there are some always-true IF clauses that can be removed.
However, I didn't do that in this patch because I just wanted to keep it
nice & small. Note that the call to getContactID() in the preferred language
function is always FALSE but in the handleUpdate() routine it
is not actually called when language is not there so there
is no need for the IF

Bug: T357471

Change-Id: I8244faa946ddb79803933830b3b7b412b11c0069