Add test listener to make test clean up more consistent
It turns out upstream Civi uses a test listener and when we
run our extension tests separately this is used.

When we run our test suite it is not - so we maybe miss
the chance to be really consistent across tests.

This adds a class based on the Civi\Test\CiviTestListener
but so cut down it is unrecognizable.

It might be a better place than our WMFEnvironmentTrait
for some things - although I'm kinda undecided atm

Bug: T370371

Change-Id: I03a50975d5d539456fe7b111530f419fa3bfd4f8