Replace a chunk of single character outputs with summary text
I tried to include summaries showing how many of each transaction type
were found & missing by payment method as that seemed to be liked.

However, there is a bit of a risk of falling down a code rabbit hole when
the rest of the output is still kinda kooky. Hopefully this is an OK compromise

Bug: T270675

Change-Id: If6f42057a500ba5578b3df4067e5e47d8f90efa7