Update deduper with fixes merged upstream
I'm trying to address a specific issue around typing which can be
replicated on staging with the drush command

drush @wmff cvapi Contact.merge mode=safe to_keep_id=460068 to_remove_id=931898

However, I have had quite the fight because the CI running on the extension
has different base settings to our WMF install & we hadn't been running it outside
WMF context & hence there was a bit to fix up. Unfortunately all this mucking
around makes for quite a brutal patch - it is broken down into
individual commits upstream as I worked through the test issues

The bug we are hitting was exposed by, but not caused by, the tightening of
php types.

The issue occurs where

where we are merging a contact with
- email address = home (primary)
- email address = billing
- email address = main

Into a contact with
- email address = home (primary)
- email address = billing

Where only the 2 home primary addresses are the same.

Note it is important here that main has a lower location type ID than
billing does. When the merge takes place the location arrays are keyed by
location type ID - meaing that in the contact to remove
the block number for the billing location is 2
whereas for the contact to keep it is 1 - the code is expecting them
to have the same block number and a recent-ish addition
of typing made a soft miss into a hard error

Bug: T363965

Change-Id: I19272d3392b3e0ee57982eb3d2e330716f4bbf0f