Improve cache-flush time out issue
I was hitting
Fatal error: Maximum execution time of 0 seconds exceeded

when doing a cache flush from the UI - it seems it might be a loop
https://stackoverflow.com/questions/4869611/fatal-error-maximum-execution-time-of-0-seconds-exceeded

Removing this api call resolved it for me. I'm not sure why but the other
one seems OK for now. This might have started happening last Civi upgrade
with some change in what hooks are called when

Change-Id: I6872c1fe0070e8b3e730ec29e2e06dce8e68b148