Add matching support for full_name field
This implements an import hook in the deduper code to map the full_name onto
the relevant name fields. I had originally thought to do a hook in the matching process
but actually it is important for our code that the full name is unpacked before
our WMF hook runs. I switched it over to use the symfony listener in order
to control the weight

https://docs.civicrm.org/dev/en/latest/hooks/

Note there is a fidelity import file in the offline2civicrm/tests/data folder
with some contacts with full_names

Bug: T373484
Change-Id: Iddf29f62a9348abab5d00cdaa49013069801c6b2