Get past the validate check
This gets us past the total_amount required for validate.

I would like to do a better fix & started some discussion to that
effect -
https://lab.civicrm.org/dev/core/-/issues/5456

However, I'm mindful that we need to get this to Melanie asap so this is enough for now.

The foreign_checks_trilogy.csv is useful for testing here

Bug: T368998

Change-Id: I236cddf965473d89117c89bdd3aa7e5bb8266565