Reapply our hack Temporary fix for handling duplicate contacts on import
This patch causes a new contact to be created when there is more than
one possible duplicate - the new contact is blank but the ID is used in the import.

The contact is added to a new group for deduping.

I was close to figuring out a hook way to do this but realise it requires a field
to be in the csv for contact ID and that is not currently the case
for the imports that Melanie is working on & with her iminent leave I think a quick
hack is called for - I will keep looking at a better fix to get
this back out of core

Bug: T374063
Change-Id: I637c5602840e5cc69b5f398f3d1355803c4b3ef7