Temporary fix for handling duplicate contacts on import
This patch causes a new contact to be created when there is more than
one possible duplicate - the new contact is blank but the ID is used in the import.

The contact is added to a new group for deduping.

I was close to figuring out a hook way to do this but realise it requires a field
to be in the csv for contact ID and that is not currently the case
for the imports that Melanie is working on & with her iminent leave I think a quick
hack is called for - I will keep looking at a better fix to get
this back out of core

For testing - try creating 2 contacts with the email
address rambo@mouse.com and then importing the engage_individual.csv
file from the wmf_civicrm/tests/data folder. You will need
to pay attention to ensuring that the contact is set to
'create or update' but can stick with the default rule.
Map ContributionType to Financial Type
Bug: T374063
Change-Id: Ib5044c8cb578f17dd88a73bd26f1ccd01080d4e8

Change-Id: I057aa923d0a3bba0f3495ed92cb5c61abbf72cc6