Add riverlea theme
From 5.80 the riverea theme will start shipping with core.

I'm adding it into the 5.79 civicrm so we can play with it - but
I'm not enabling on dev / prod at this point.

Specifically there is a theming difference between the riverlea
implementation of shoreditch and the theisland implementation
of shoreditch that is relevant to the user interface usability
on the batch data entry screen
https://github.com/civicrm/civicrm-core/pull/31433

There are a few ways this could go
1) that PR gets approved & merged - if however, it turns
out that it would be 'wrong' to skip tabbing to the spanners
it needs to be fixed in the theme by either
2) switching to riverlea or
3) fixing the island

Bug: T371543
Change-Id: Id069bab37aba7a21bcc42e0787c09c939acc582d