Do not include annual donors in end of year emails.
The MakeJob function determines which donors to send end of year emails to.

These are sent to donors who make monthly donations, in lieue of receipting
every single one. Annual donors would not get this as they would be receipted each
email. The exception is when they are ALSO a monthly donors, because the
receipts we sent to monthly donors include ALL donations by the given contact

Bug: T379751
Change-Id: I1c186e9ff128f794eca9112a96b3ec5ce2ded15a