Update Thankyou.send action to require less parameters
This is part of a code consolidation in our email sending towards
attachment support.

Before this patch the Thankyou.render action works with just a
contribution ID but the Thankyou.send action requires both.

This means that we wind up having to call wrapper code (apiv4 thank_you.send,
thank_you_for_contribution) in order to call the Thankyou.send.

This patch makes the api cope if not called via the wrapper code
and adds test cover

Bug: T299096
Change-Id: Ia8bca299a750bd71937bb5afd2717eab719853e6