Remove WMFException type that is partially-used in outgoing code
The offline2civicrm code just catches the exception & prints the error without
looking at the error code. There is no other use of this type and
the offline2civicrm code is up for iminent removal.

I put in the same number as I had to put something in but it is unused

Bug: T270677
Change-Id: I11895e6500f5000c1796958896e101070ff89b47