Verify sending recurring Paypal payments to the donations queue.
This test suite tries to verify that PayPal recurring payment
charges can be sent to the donations queue.

Currently, the code can't handle donations without a valid subscription
ID, which is caught by the failure of
testRecurringPaymentPaypalMissingPredecessor - which is currently skipped with:
- $this->markTestIncomplete("this one is failing due to the issue documented
here https://phabricator.wikimedia.org/T240581#9353558");

We should use that test to help confirm the fix once it's in place.

Bug: T366526
Change-Id: Iea765a71aa9b9a898f96bd61bd207434922bcb12