Populate new snooze_end_date 'real field'
Acoustic has a system field internally called 'snoozeEndData' - this
field is largely hidden in Acoustic - although we do display it on (e.g)
https://wmff.localhost:32353/civicrm/a/#/omnimail/remote-contact?cid=203

The field has several limitations - Acoustic logic says it
should be unset if 'remind-me-later' is initiated, if can't
be searched on and it can't be used in 'decision gates' in the
remind-me-later flow.

To get around this we have determined we should push the snooze
date to BOTH the system snooze date AND a new 'normal' field snooze_end_date.
Note that this patch shows the new field as the snooze_end_date
- this might cause a flood of attempts to update the snooze date
in Acoustic using the verifySnooze job - this is mostly a good
thing as it will serve to back fill but I think it
makes sense to disable the job while rolling out

Bug: T386023
Change-Id: I9ded871547f08ebdbb8c72acb81d335132256fbe