Stop putting REPLY MAIL BLOCK emails on hold
Recent discussion with Brian on the phab has clarified that
this is a mis-interpretation of REPLY MAIL BLOCK that
we implemented right at the beginning. REPLY MAIL BLOCK
is likely transitory and should not be put on hold in Civi.

We need to consider data clean up next but this will stop us
putting more on hold

Bug: T383193
Change-Id: I3b13b94f3de97177181da143acf4006c032f3233