More CRM_Core_Exception error handling up to handleError()
This is intended to be the start of moving error handling up the chain in a handful of places. The logic is
that when a CRM_Core_Exception is thrown there are 3 possibilities

1) The exception is expected in the code (e.g. a duplicate contribution) and there is specifically handled
2) The exception is deadlock related
3) The exception is some other unplanned for Civi-error

In the case of the first it makes sense for that exception to be caught & handled in the calling code. However,
for the other 2 the calling code just catches it to, inconsistently, convert it an appropriate
WMF Exception because that is all the upstream understands. If if it not handled (by design or
by ommission) it blocks further queue processing rather than pushing to the damaged queue.

This fixes it such that we no longer need to catch categories 2 & 3 in the code (there are
multiple other places, eg. in the RecurringDonation class where we could remove
deadlock handling with this merged).

The one 'hmmm' is whether there is a case for crashing rather than pushing to the damaged
queue when unexpected damaged items occur. I lean to 'no' because it would only make sense if
there were tonnes of them - ie a systematic error - but that feels uncommon & requeuing
from damaged is OK there I think....

Note the test I added today steps through this - testDuplicateTrxnIDHandling()

Bug: T365418
Change-Id: Ifa91983a2730f3ccbc7929f84764ee74e557c095