Fixes to notifications on queueErrors
This ports 2 upstream PRs
https://github.com/civicrm/civicrm-core/pull/32367
https://github.com/civicrm/civicrm-core/pull/32365

& makes one local tweak. It arises from concerns that we are finding out
too little too late about coworker fails. Ie we only get notified when coworker fully
gives up and we don't get the original error message. The 2 core patches
address the content of the message.

The other issue is we only get notified on the last fail - which is a pretty
drawn out feedback mechanism. This fixes to do on the first & last

Bug: T366635

Change-Id: I21047721a5b839bdfc869efa9977a7acfb2ccfe2