Update Fidelity Import
This adds Fidelity specific handling, along with test fixes to make it testable
per discussions with Melanie. Note that this requires work & testing by Melanie
to make it operational so it will not be immediately live on deployment.

Details
1) For Fidelity imports we add a second soft credit to the Fidelity Charitable Gift Fund
with the soft credit type 'Banking Institution' in the import hook
2) For the Fidelity imports they only provide 1 address column so we copy the address
from the main contact to the soft credit contact, in the import hook
3) In general we have decided to create a Donor Advised relationship where soft credits
of the donor-advised_fund are made. This is in the ContributionSoft pre hook as
it is not restricted to imports.
4) In order to test the last item I had to do some test fixes to alter the date
in the fidelity.csv as it needs to be recent to trigger the hook
5) I tweaked the details in fidelity.csv to make sure we were deleting
the individuals and that we were testing an attempt to soft credit where the
relationship already exists

I'm working on docs at
https://wikitech.wikimedia.org/w/index.php?title=Fundraising/Internal-facing/CiviCRM/Imports&veaction=edit
Bug: T386017

Change-Id: Iba737d371f0b9fd087126baf6c7d942a0579cae2