Stop recurring failmail when making API calls
We added the failmail alerts to catch the case when we rejected our
own payment attempt in our validation code. We've fixed most of those
cases and are now getting a lot of failmail about normal failures
from the processor APIs. Those can just go in the log like payment
errors.

Change-Id: I9ef35fdd3f2d6770f7da932fc1df5524c34818b9