Remove drush/drush from composer
The php-http/discovery plugin was not loaded as it is not listed in allow-plugins and is not required by the root package anymore.
./composer.json has been updated
The php-http/discovery plugin was not loaded as it is not listed in allow-plugins and is not required by the root package anymore.
Running composer update drush/drush --with-all-dependencies
Loading composer repositories with package information

Running composer update drush/drush --with-all-dependencies
Loading composer repositories with package information                                                               Updating dependencies
Lock file operations: 0 installs, 0 updates, 13 removals
  - Removing consolidation/annotated-command (2.12.2)
  - Removing consolidation/output-formatters (3.5.1)
  - Removing drush/drush (8.5.0)
  - Removing pear/console_table (v1.3.1)
  - Removing psy/psysh (v0.12.9)
  - Removing symfony/console (v4.4.49)
  - Removing symfony/finder (v4.4.44)
  - Removing symfony/polyfill-php73 (v1.32.0)
  - Removing symfony/process (v4.4.44)
  - Removing symfony/var-dumper (v5.4.48)
  - Removing webflo/drupal-finder (1.3.1)
  - Removing webmozart/assert (1.11.0)
  - Removing webmozart/path-util (2.3.0)

Change-Id: I818ba4a9f5de555be59d07904284578fedbb2cee