Use apiv4 in updateWMFDonorLastDonation
If we do the following on prod
grep  "is changing values" thank_you_mail_send_smtp_frmx2001/*

We get a LOT of rows. I suspect the comparison is off but the
from & to are in different formats ie

"from":{"amount":6.5300000000000002,"date":"2025-07-24 01:24:09","amount_usd":6.5300000000000002,"currency":"USD"},
"to":{"custom_129":"AUD","custom_130":"10.00"}}

This updates to apiv4 & aligns everything to refer to the actual field names.

Bug: T400309
Change-Id: Ibd41d89b552ccfc02543192c4f0aaa897639dd4f